Remove arel API
