--- spec/arel_integration_spec.rb
@@ -1,59 +1,0 @@
-# frozen_string_literal: true
-
-require "spec_helper"
-
-RSpec.describe "ArelIntegrationTest" do
-  before do
-    @builder = ParadeDB::Arel::Builder.new(:products)
-  end
-
-  def sql(node)
-    ParadeDB::Arel.to_sql(node)
-  end
-  it "full matrix of operators" do
-    nodes = []
-    nodes << @builder.match(:description, "running shoes")
-    nodes << @builder.match_any(:description, "wireless bluetooth")
-    nodes << @builder.phrase(:description, ParadeDB.slop("running shoes", 2))
-    nodes << @builder.term(:description, ParadeDB.boost(ParadeDB.fuzzy("shose", 2, prefix: false), 2))
-    nodes << @builder.term(:description, "literal")
-    nodes << @builder.term_set(:category, %w[audio footwear])
-    nodes << @builder.regex(:description, "run.*")
-    nodes << @builder.near(:description, ParadeDB.proximity("sleek").within(1, "shoes"))
-    nodes << @builder.phrase_prefix(:description, "run", "sh")
-    nodes << @builder.more_like_this(:id, 5, fields: [:description, :category])
-    nodes << @builder.full_text(:description, "pdb.all()")
-
-    rendered = nodes.map { |n| sql(n) }
-
-    assert_includes rendered, %("products"."description" &&& 'running shoes')
-    assert_includes rendered, %("products"."description" ||| 'wireless bluetooth')
-    assert_includes rendered, %("products"."description" ### 'running shoes'::pdb.slop(2))
-    assert_includes rendered, %("products"."description" === 'shose'::pdb.fuzzy(2)::pdb.boost(2))
-    assert_includes rendered, %("products"."description" === 'literal')
-    assert_includes rendered, %("products"."category" @@@ pdb.term_set(ARRAY['audio', 'footwear']))
-    assert_includes rendered, %("products"."description" @@@ pdb.regex('run.*'))
-    assert_includes rendered, %("products"."description" @@@ ('sleek' ## 1 ## 'shoes'))
-    assert_includes rendered, %("products"."description" @@@ pdb.phrase_prefix(ARRAY['run', 'sh']))
-    assert_includes rendered, %("products"."id" @@@ pdb.more_like_this(5, ARRAY['description', 'category']))
-    assert_includes rendered, %("products"."description" @@@ pdb.all())
-  end
-  it "boolean chains" do
-    base = @builder.match(:description, "running").and(
-      @builder.phrase(:description, "trail shoes").not
-    )
-
-    other = @builder.match_any(:category, "Footwear").and(
-      @builder.term(:in_stock, true)
-    )
-
-    combined = base.or(other)
-    rendered = sql(combined)
-    assert_includes rendered, "&&& 'running'"
-    assert_includes rendered, "### 'trail shoes'"
-    assert_includes rendered, "||| 'Footwear'"
-    assert_includes rendered, "=== TRUE"
-    assert_includes rendered, "AND NOT"
-    assert_includes rendered, " OR "
-  end
-end
